Fixed dependencies. Added bootstrap.
This has Chris's code, and it also has everything working the way it's supposed to. The master schedule populates good, the code can be run with npm start, and the server can be run by navigating to ./src/server/index.js and using node index.js in a separate terminal window.