massive update
tables are now populate with the following hierarchy:
home (where the navbar is) > query handler > display (where make table happens)

events bubble up from the click => query handler gets data, stores it, pass it in a property function defined within display. i had to use setState on the argument and update an array within display in order to get this to work, otherwise my .map wouldn't work and would say it was undefined :(

FOR TESTING: Right now, this works with STUDENT. Each role will get their own query handler and home, but they will all route to the same display. Queries need to be customized based on whatever that query is (as arguments can be different) inside the routes.

TO TEST POPULATING TABLE:
Load project, on nav bar click "Course Search" to generate MASTER SCHEDULE (i was testing, hence the mismatch) and unofficial transcript to get TRANSCRIPT (no student id needed, as it is predefined). Both are two separate queries and will populate the table. To continue this, implement Home & Query handler files and make sure to give home objects properties representing the query handler methods. Query handler then makes a call to display, which again, all user types should share.

Routes need to be configured as well, as in, the logic needs to be inserted. that's all for now!