trying to figure out input fields
create student has a name not really appropriate for what it does, it's actually a work in progress and is meant to display all of the user inputs. well, it's repurposed - that originally wasn't the purpose, but here we are. dynamicforms is exactly as it sounds, but instantiating it didn't really appear to actually make forms appear for some reason, but i'll leave it there in case a solution comes up later. basically whenever "makefields" appears it was likely me testing something. the neason why the navbar component gives a 'field' property to each object is because i was gonna read from the fields object to determine which fields to display. couldn't quite get it working tho