Request bodies are dynamically created, & more
Request bodies are generated in the query handler, using a needed [] array with a list of properties particular to that query, it will automatically fill in any empty fields with a null string"".
Insert queries now dynamically replace ? tokens with vals. Accomplished this w/ replaceQuestionMarks in admin routes.

These are import changes b/c the groundwork is set for applying this to the other queries. Now we need to make sure every button works, that is next on the agenda.